watcher/requirements.txt
glfw
PyOpenGl
pyrr
freetype-py
numpy



